+++ tests/integration/test_otlp_logs.py
@@ -35,6 +35,10 @@ def test_otlp_logs_conversion(
         "organizations:ourlogs-ingestion",
         "organizations:relay-otel-logs-endpoint",
     ]
+    project_config["config"]["retentions"] = {
+        "log": {"standard": 30, "downsampled": 13 * 30},
+    }
+
     relay = relay(relay_with_processing(options=TEST_CONFIG), options=TEST_CONFIG)
 
     ts = datetime.now(timezone.utc)
@@ -135,7 +139,8 @@ def test_otlp_logs_conversion(
                 "sentry.observed_timestamp_nanos": {
                     "stringValue": time_within(ts, expect_resolution="ns")
                 },
-                "sentry.payload_size_bytes": {"intValue": "358"},
+                "sentry.origin": {"stringValue": "auto.otlp.logs"},
+                "sentry.payload_size_bytes": {"intValue": "385"},
                 "sentry.severity_text": {"stringValue": "info"},
                 "sentry.span_id": {"stringValue": "eee19b7ec3c1b174"},
                 "sentry.timestamp_nanos": {
@@ -185,7 +190,7 @@ def test_otlp_logs_conversion(
             "org_id": 1,
             "outcome": 0,
             "project_id": 42,
-            "quantity": 358,
+            "quantity": 385,
         },
     ]
 
@@ -202,6 +207,10 @@ def test_otlp_logs_multiple_records(
         "organizations:ourlogs-ingestion",
         "organizations:relay-otel-logs-endpoint",
     ]
+    project_config["config"]["retentions"] = {
+        "log": {"standard": 30, "downsampled": 13 * 30},
+    }
+
     relay = relay(relay_with_processing(options=TEST_CONFIG), options=TEST_CONFIG)
 
     ts = datetime.now(timezone.utc)
@@ -249,6 +258,7 @@ def test_otlp_logs_multiple_records(
                 "sentry.observed_timestamp_nanos": {
                     "stringValue": time_within(ts, expect_resolution="ns")
                 },
+                "sentry.origin": {"stringValue": "auto.otlp.logs"},
                 "sentry.payload_size_bytes": {"intValue": mock.ANY},
                 "sentry.severity_text": {"stringValue": "error"},
                 "sentry.span_id": {"stringValue": "eee19b7ec3c1b174"},
@@ -288,6 +298,7 @@ def test_otlp_logs_multiple_records(
                 "sentry.observed_timestamp_nanos": {
                     "stringValue": time_within(ts, expect_resolution="ns")
                 },
+                "sentry.origin": {"stringValue": "auto.otlp.logs"},
                 "sentry.payload_size_bytes": {"intValue": mock.ANY},
                 "sentry.severity_text": {"stringValue": "debug"},
                 "sentry.span_id": {"stringValue": "eee19b7ec3c1b175"},
@@ -337,6 +348,6 @@ def test_otlp_logs_multiple_records(
             "org_id": 1,
             "outcome": 0,
             "project_id": 42,
-            "quantity": 251,
+            "quantity": 305,
         },
     ]